modified:   datasets/Dataset_cls.py 	modified:   datasets/__pycache__/Dataset_cls.cpython-38.pyc 	modified:   main.py 	modified:   models/__pycache__/backbone.cpython-38.pyc 	modified:   models/__pycache__/detr.cpython-38.pyc 	modified:   models/backbone.py 	modified:   models/detr.py 	modified:   util/__pycache__/misc.cpython-38.pyc 	modified:   util/__pycache__/train_utils.cpython-38.pyc 	modified:   util/misc.py 	modified:   util/train_utils.py

+++ util/misc.py
@@ -289,7 +289,7 @@ class NestedTensor(object):
 
     def to(self, device):
         # type: (Device) -> NestedTensor # noqa
-        cast_tensor = self.tensors.to(device).float()
+        cast_tensor = self.tensors.to(device)
         mask = self.mask
         if mask is not None:
             assert mask is not None
